feat(message.ts): merge message.ts for background and content scripts
