+++ src/background/adguardEngine/index.ts
@@ -3,7 +3,7 @@ import * as utils from '../../core/utils'
 import * as logger from '../../core/logger'
 import * as http from '../../core/http'
 import * as store from '../../core/store'
-import * as message from '../message'
+import message from '../../core/message'
 import * as statistics from '../statistics'
 import $ from 'jquery'
 import IAdblockEngine from '../../adblockEngine/adblockerEngine'
