+++ src/background/download.ts
@@ -2,7 +2,7 @@ import * as coreConfig from '../core/config'
 import * as grade from '../core/grade'
 import * as notification from '../core/notification'
 import * as i18n from '../core/i18n'
-import * as message from './message'
+import message from '../core/message'
 
 let downloadSafetyEnabled: boolean = false
