+++ src/content/block.ts
@@ -1,6 +1,5 @@
 import * as coreConfig from '../core/config'
 import * as utils from '../core/utils'
-import * as message from '../core/message'
 import $ from 'jquery'
 
 $.ready.then(() => {
